fix(rythmo): draw on the real character advance, not a rounded cell
The grid cell was the advance rounded to a whole pixel, 13 px against
the 12.59 px the default font actually advances. Every character was
pushed 0.41 px further than the glyph wanted, so the band came out 3.2%
wider than before and the text read looser and softer.

charWidth() keeps the fractional advance. Measured against the previous
rendering, at the four sub-pixel phases of a scrolling band: 458 pixels
out of 212800 differ by more than 25 levels, solid-ink ratio 38.1-38.7%
against 38.0-38.6%, mean grey within 0.2. The cells still hold their
time grid: the offset between a character and its slot stays constant
from index 0 to 5000, at 1280, 1920 and 2560 px wide.

+++ src/gui/RythmoWidget.h
@@ -125,7 +125,7 @@ protected:
 private:
   // Helpers
   bool isDarkTheme();
-  int charWidth() const;
+  double charWidth() const;
   double nominalCharMs() const;
   double pixelsPerMs() const;
   int cursorIndex() const;
@@ -154,7 +154,7 @@ private:
   int m_lastMouseX;
 
   // Font cache
-  mutable int m_cachedCharWidth;
+  mutable double m_cachedCharWidth;
 
   // Seek debouncing
   QTimer *m_seekTimer;
